requirements.txt
scrapy
scrapy-selenium
pika
PyYAML